Add portable CaloRecHitSoA data format
This data format will serve as the base for ParticleFlow RecHit construction using Alpaka. It contains only those members necessary ParticleFlow.
This is a temporary solution. As soon as official Alpaka-compatible SoAs become available from the HCAL/ECAL groups, this format becomes obsolete and will be replaced by those. This format also serves as a base for these two groups. The official formats should be super-sets of this one.