Add ESProducer to supply cut parameters to Alpaka ParticleFlow RecHit producer
Calorimeter hits are subjected to certain quality cuts to decide if they should become ParticleFlow RecHits. This commit adds a data structure to hold the parameters for these cuts and two modules that fill this data structure. Currently, this module only supports a subset of the types quality cuts that are supported by the CPU backend. Further relevant types of cuts still need to be identified and implemented.