Add conversion from PFRecHitSoA to legacy PFRecHit collection
This commit adds a producer that reads the PFRecHitSoA from the Alpaka module and produces a PFRecHit collection in the same format as the CPU PFRecHitProducer. This conversion is necessary as long as not all consumers of the PFRecHits are updated to use the portable SoA format. It can also be used to validate the Alpaka implementation of the PFRecHitProducer.