Prepare for Alpaka implementation of calorimeter reconstruction
The HCAL and ECAL groups will use different SoAs, when they construct their calorimeter hits using Alpaka. So far, the PFRecHitProducer uses the CaloRecHitSoA type everywhere, which was introduced as a temporary solution, as long as no official SoAs are available. This commit makes the SoA type configurable separately for HCAL and ECAL. This should allow for an easy transition to SoAs, once they become available.